Update sieve.asm
Sorry for not adding this until now, my family wanted to spend time with me. I did most of the same things you did from what I can tell(probably a bit more sloppy as I didn't comment out a lot of the original code I changed and just deleted it), and I also changed a bit of stuff in init. I have it doing a beq for $t0 and $t9 to go to a separate label to change $t0's value, which gives the whole thing a 17.2% increase, as opposed to an increase of 13.8%. I wasn't sure if I should commit to the master branch, but I figured to make a different branch.