22.05% optimization
I dunno why the adding outer by 2 causes an error, but when I was working on it last night I noticed that if you leave that addi in outer to 2 and set the number that $t0 increments by in count: to 2 that it won't print off those non prime powers of two, have an optimization of around 57%, but it doesn't print 2. I'll look more into that today, but I figured maybe if I let you know about the count section that you could see what is going on with it not printing 2.