Update sieve.asm
moved the mfhi to below the bgt, it saves only 24 total cycles though, making this 37.28% optimized instead of the about 37.21% it was at. I'm not quite sure how to eliminate the cache misses in sw in inner and I really don't know where else we could optimize it other than in the cache.